Refactor and update licensing in Program.cs
- Updated copyright notice to 2024 and switched to Apache License 2.0.
- Introduced async Task RunSample() for better organization.
- Encapsulated engine creation and logon within RunSample.
- Shifted request handling to OnRequestReceived for modularity.
- Consolidated Request and Response classes into Program.cs.
- Updated C# language version to 9 in .csproj and removed old references.
